agent/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
boto3>=1.34.0
requests>=2.31.0
PyJWT>=2.8.0
cryptography>=41.0.0
pydantic>=2.5.0
httpx>=0.26.0
